requirements.txt
alabaster==0.7.16
anyio==4.3.0
asgiref==3.8.1
Babel==2.14.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.4
django-filter==24.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
docutils==0.21.2
h11==0.14.0
idna==3.7
imagesize==1.4.1
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.0
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
snowballstemmer==2.2.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
Sphinx==7.3.7
sphinx-autobuild==2024.4.16
sphinxcontrib-applehelp==1.0.8
sphinxcontrib-devhelp==1.0.6
sphinxcontrib-htmlhelp==2.0.5
sphinxcontrib-httpdomain==1.8.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.7
sphinxcontrib-serializinghtml==1.1.10
sqlparse==0.5.0
starlette==0.37.2
urllib3==2.2.1
uvicorn==0.29.0
watchfiles==0.21.0
websockets==12.0
